requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
chardet==5.2.0
click==8.1.8
cloudpickle==3.1.1
cobble==0.1.4
colorama==0.4.6
colorlog==6.9.0
cssselect2==0.8.0
datasets==3.4.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.5.41
EbookLib==0.18
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
google-auth==2.38.0
google-genai==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keyring==25.6.0
cryptography==46.0.3
json_repair==0.39.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.51.0
loguru==0.7.3
lxml==5.3.1
magicattr==0.1.6
Mako==1.3.9
mammoth==1.9.0
markdown2==2.5.3
markdown_analysis==0.1.4
markdownify==0.13.1
marker-pdf==1.6.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.31.0
networkx==3.4.2
numpy==1.24.3
openai==1.66.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optuna==4.2.1
packaging==24.2
pandas==2.1.0
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.6
plotly==5.17.0
dash==2.14.0
dash-bootstrap-components==1.4.2
propcache==0.3.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydyf==0.11.0
pylatexenc==2.10
pypdfium2==4.30.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
surya-ocr==0.13.0
sympy==1.13.1
tenacity==9.0.0
textdistance==4.6.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.15.0
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
websockets==14.2
win32_setctime==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zopfli==0.2.3.post1
